Add force flag and move tagging updated logic to FindVaccineCenter
